% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess-package.R
\docType{package}
\name{ess-package}
\alias{ess}
\alias{ess-package}
\title{ess: Download Data from the European Social Survey on the Fly}
\description{
Download data from the European Social Survey directly from their website <http://www.europeansocialsurvey.org/>. There are two family of functions that allow you to download and interactively check all countries and rounds available.
}
\details{
For detailed examples visit the package
 website at \url{https://cimentadaj.github.io/ess/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cimentadaj/ess}
  \item Report bugs at \url{https://github.com/cimentadaj/ess/issues}
}

}
\author{
\strong{Maintainer}: Jorge Cimentada \email{cimentadaj@gmail.com}

}
\keyword{internal}
