% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_all_cntrounds.R
\name{ess_all_cntrounds}
\alias{ess_all_cntrounds}
\title{Download all available integrated rounds for a country from the European Social Survey}
\usage{
ess_all_cntrounds(country, your_email, output_dir = ".",
  only_download = FALSE)
}
\arguments{
\item{country}{A character of length 1 with the full name of the country. Use show_countries()
for a list of available countries.}

\item{your_email}{a character vector with your email, such as "your_email@email.com".}

\item{output_dir}{a character vector with the output directory in case you want to only download the files using
the \code{only_download} argument. Defaults to the current working directory. Files will be saved
as ESS_*/ESS\* where the first star is the country name and the second start the round number.}

\item{only_download}{whether to only download the files as Stata files. Defaults to FALSE.}
}
\value{
if \code{only_download} is set to FALSE it returns a list containing the latest
version of all available rounds. If \code{only_download} is set to TRUE, it returns nothing but
saves all the rounds in .dta format in \code{output_dir}
}
\description{
Download all available integrated rounds for a country from the European Social Survey
}
\examples{
\dontrun{

# Will return all available rounds for Denmark
ess_all_cntwaves("Denmark", "your_email@gmail.com")

# Will download all waves to directory "./denmark/
# as .dta files and won't return the rounds in R.
ess_all_cntwaves("Denmark", "your_email@gmail.com", output_dir = "./denmark/", only_download = TRUE)

}
}
