% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_email.R
\name{set_email}
\alias{set_email}
\title{Save your ESS email as an environment variable}
\usage{
set_email(ess_email)
}
\arguments{
\item{ess_email}{a character string with your registered email.}
}
\description{
Save your ESS email as an environment variable
}
\details{
You should only run \code{set_email()} once and every \code{ess_*} function
should work fine. Make sure your email is registered at 
\url{http://www.europeansocialsurvey.org/} before setting the email.
}
\examples{

\dontrun{
set_email("my_registered@email.com")

ess_country(1)
}

}
