% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_funs.R
\name{show_countries}
\alias{show_countries}
\title{Return available countries in the European Social Survey}
\usage{
show_countries()
}
\value{
character vector with available countries
}
\description{
Return available countries in the European Social Survey
}
\examples{

show_countries()

}
