% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_any_rounds.R
\name{show_country_rounds}
\alias{show_country_rounds}
\title{Return available rounds for a country in the European Social Survey}
\usage{
show_country_rounds(country)
}
\arguments{
\item{country}{A character of length 1 with the full name of the country.
Use \code{\link{show_countries}}for a list of available countries.}
}
\value{
numeric vector with available rounds for \code{country}
}
\description{
Return available rounds for a country in the European Social Survey
}
\examples{

show_country_rounds("Spain")

show_country_rounds("Turkey")

}
