% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_country_waves.R
\name{show_country_waves}
\alias{show_country_waves}
\title{Helper function to return available waves for a country in the European Social Survey}
\usage{
show_country_waves(country)
}
\arguments{
\item{country}{A character of length 1 with the full name of the country. Use show_countries()
for a list of available countries.}
}
\value{
character vector with available waves for \code{country}
}
\description{
Helper function to return available waves for a country in the European Social Survey
}
\examples{

show_country_waves("Germany")

show_country_waves("Spain")

show_country_waves("Turkey")

}
