% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rounds.R
\name{show_rounds}
\alias{show_rounds}
\title{Helper function to return available rounds in the European Social Survey}
\usage{
show_rounds()
}
\value{
numeric vector with available rounds
}
\description{
Helper function to return available rounds in the European Social Survey
}
\examples{

show_rounds()

}
