% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sddf_rounds.R
\name{show_sddf_rounds}
\alias{show_sddf_rounds}
\title{Return available SDDF rounds for a country in the European Social Survey}
\usage{
show_sddf_rounds(country)
}
\arguments{
\item{country}{A character of length 1 with the full name of the country.
Use \code{\link{show_countries}}for a list of available countries.

@details SDDF data are the equivalent weight data used to analyze the European
Social Survey properly.}
}
\value{
numeric vector with available rounds for \code{country}
}
\description{
Return available SDDF rounds for a country in the European Social Survey
}
\details{
Given that the new SDDF files from the ESS are now comming in a fully integrated
file rather than separated by countries, this function only checks for SDDF files in
rounds 1 through 6.
}
\examples{

show_sddf_rounds("Spain")

}
