% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_funs.R
\name{show_themes}
\alias{show_themes}
\title{Return available themes in the European Social Survey}
\usage{
show_themes()
}
\value{
character vector with available themes
}
\description{
Return available themes in the European Social Survey
}
\examples{

show_themes()

}
