% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_waves.R
\name{show_waves}
\alias{show_waves}
\title{Helper function to return available waves in the European Social Survey}
\usage{
show_waves()
}
\value{
numeric vector with available waves
}
\description{
Helper function to return available waves in the European Social Survey
}
\examples{

show_waves()

}
