% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_funs.R
\name{show_rounds}
\alias{show_rounds}
\title{Return available rounds in the European Social Survey}
\usage{
show_rounds()
}
\value{
numeric vector with available rounds
}
\description{
Return available rounds in the European Social Survey
}
\examples{

\dontrun{
show_rounds()
}


}
